changed main.h to struct.h and add struct.c file
